requirements.txt
python-dotenv
python-telegram-bot
Appium-Python-Client
beautifulsoup4
JPype1
selenium